requirements.txt
flask==3.1.0
openpyxl==3.1.5
anthropic>=0.80.0
python-docx>=1.1.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
python-dotenv>=1.0.0
